bug fix-Bug 4531

See merge request almullagroup/amx-jax!569

+++ ms-jax/src/main/java/com/amx/jax/manager/RemittanceTransactionManager.java
@@ -291,7 +291,7 @@ public class RemittanceTransactionManager {
 
 	private void validateRiskyBene(BenificiaryListView beneficiary, Customer customer) {
 		if (jaxConfigService.getBooleanConfigValue(JaxDbConfig.BLOCK_BENE_RISK_TRANSACTION, true)) {
-			if (beneficiary.getCountryId() != customer.getNationalityId()) {
+			if (beneficiary.getCountryId().intValue() != customer.getNationalityId().intValue()) {
 				int beneCountryRisk = countryService.getCountryMaster(beneficiary.getCountryId()).getBeneCountryRisk();
 				if (beneCountryRisk == 1) {
 					throw new GlobalException(JaxError.BENE_COUNTRY_RISK, "Bene country risk");
